+++ core/src/main/java/com/crowdar/core/LocatorManager.java
@@ -6,7 +6,8 @@
 package com.crowdar.core;
 
 import com.crowdar.core.pageObjects.LocatorTypesEnum;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.By;
 
 
@@ -17,7 +18,7 @@ public class LocatorManager {
             String[] locatorProperty = locatorElement.split(":");
             return getLocatorInEnum(locatorProperty, locatorReplacementArgs);
         } catch (NullPointerException e){
-            Logger.getLogger(LocatorManager.class).error(e.getMessage());
+            LogManager.getLogger(LocatorManager.class).error(e.getMessage());
             throw new RuntimeException(String.format("Locator property %s was not found", locatorElement));
         }
     }
@@ -29,7 +30,7 @@ public class LocatorManager {
             String value = String.format(locatorProperty[1], locatorReplacementArgs);
             return LocatorTypesEnum.get(type).getLocator(value);
         } catch (IndexOutOfBoundsException e) {
-            Logger.getLogger(LocatorManager.class).error(e.getMessage());
+            LogManager.getLogger(LocatorManager.class).error(e.getMessage());
             throw new RuntimeException("Locator property format is invalid. Example: css:#loginButton");
         }
     }