+++ core/src/main/java/com/crowdar/email/EmailPropertiesEnum.java
@@ -1,7 +1,8 @@
 package com.crowdar.email;
 
 import com.crowdar.core.JsonUtils;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.json.JSONObject;
 import org.testng.Assert;
 
@@ -50,7 +51,7 @@ public enum EmailPropertiesEnum {
         try {
             json = JsonUtils.getJSON(path);
         } catch (IOException e) {
-            Logger.getLogger(this.getClass()).error(e.getMessage());
+            LogManager.getLogger(this.getClass()).error(e.getMessage());
             Assert.fail(e.getMessage());
         }
         return new JSONObject(json);